refactor(ui): split the monolithic api.ts into per-domain modules
api.ts was a single 1582-line module with 159 exports. Split it into a
ui/src/api/ package behind a barrel, so every existing import from "../api"
keeps resolving unchanged:

- client.ts — shared plumbing (API_BASE, token state, fetchJSON/fetchPublic,
  401-refresh + 429-retry, getAccessToken)
- one module per domain: system, auth, zones, equipments, recipes,
  integrations, modes, history, mqtt, notifications, dashboard, energy
- index.ts re-exports the whole public surface

Pure reorganisation, no behaviour change. Public export parity verified (no
export dropped); every module is under ~235 lines.

Part of the v1.43.0 hardening batch. Closes #455
