+++ ui/src/api/index.ts
@@ -1,0 +1,14 @@
+// Barrel: re-exports the whole API surface so `import { x } from "../api"` keeps working.
+export { setAccessToken, setOnUnauthorized, getAccessToken } from "./client";
+export * from "./system";
+export * from "./auth";
+export * from "./zones";
+export * from "./equipments";
+export * from "./recipes";
+export * from "./integrations";
+export * from "./modes";
+export * from "./history";
+export * from "./mqtt";
+export * from "./notifications";
+export * from "./dashboard";
+export * from "./energy";
